Added support for transformation from XUnit tests to CppUnitTests. Platform.Collections.Methods.Console is not needed anymore.
