+++ Platform.Collections.Methods.sln
@@ -5,13 +5,6 @@ VisualStudioVersion = 16.0.29102.190
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Platform.Collections.Methods.Tests", "Platform.Collections.Methods.Tests\Platform.Collections.Methods.Tests.csproj", "{D5667200-FFE1-4D69-9124-FF8FCDC3E1AB}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Platform.Collections.Methods.Console", "Platform.Collections.Methods.Console\Platform.Collections.Methods.Console.vcxproj", "{CC13FF8A-9B80-4A09-97CF-9624DD4AED2D}"
-	ProjectSection(ProjectDependencies) = postProject
-		{D5667200-FFE1-4D69-9124-FF8FCDC3E1AB} = {D5667200-FFE1-4D69-9124-FF8FCDC3E1AB}
-		{A0967E25-55C0-465E-86A5-04067A1B7B45} = {A0967E25-55C0-465E-86A5-04067A1B7B45}
-		{4987038E-38B7-4222-8427-C5A978B9FE58} = {4987038E-38B7-4222-8427-C5A978B9FE58}
-	EndProjectSection
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CSharpToCppTranslator", "CSharpToCppTranslator\CSharpToCppTranslator.csproj", "{A0967E25-55C0-465E-86A5-04067A1B7B45}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Platform.Collections.Methods", "Platform.Collections.Methods\Platform.Collections.Methods.csproj", "{4987038E-38B7-4222-8427-C5A978B9FE58}"
@@ -22,6 +15,9 @@ EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Platform.Collections.Methods.TemplateLibrary", "cpp\Platform.Collections.Methods\Platform.Collections.Methods[cpp].vcxitems", "{5F8D8B8C-A6A0-41EF-B127-98EEDD0773E3}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Platform.Collections.Methods.TemplateLibrary.Tests", "cpp\Platform.Collections.Methods.Tests\Platform.Collections.Methods.TemplateLibrary.Tests.vcxproj", "{A29D2221-D4BA-4CD0-9D71-0CFDF33C7582}"
+	ProjectSection(ProjectDependencies) = postProject
+		{D5667200-FFE1-4D69-9124-FF8FCDC3E1AB} = {D5667200-FFE1-4D69-9124-FF8FCDC3E1AB}
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -45,16 +41,6 @@ Global
 		{D5667200-FFE1-4D69-9124-FF8FCDC3E1AB}.Release|x64.Build.0 = Release|Any CPU
 		{D5667200-FFE1-4D69-9124-FF8FCDC3E1AB}.Release|x86.ActiveCfg = Release|Any CPU
 		{D5667200-FFE1-4D69-9124-FF8FCDC3E1AB}.Release|x86.Build.0 = Release|Any CPU
-		{CC13FF8A-9B80-4A09-97CF-9624DD4AED2D}.Debug|Any CPU.ActiveCfg = Debug|Win32
-		{CC13FF8A-9B80-4A09-97CF-9624DD4AED2D}.Debug|x64.ActiveCfg = Debug|x64
-		{CC13FF8A-9B80-4A09-97CF-9624DD4AED2D}.Debug|x64.Build.0 = Debug|x64
-		{CC13FF8A-9B80-4A09-97CF-9624DD4AED2D}.Debug|x86.ActiveCfg = Debug|Win32
-		{CC13FF8A-9B80-4A09-97CF-9624DD4AED2D}.Debug|x86.Build.0 = Debug|Win32
-		{CC13FF8A-9B80-4A09-97CF-9624DD4AED2D}.Release|Any CPU.ActiveCfg = Release|x64
-		{CC13FF8A-9B80-4A09-97CF-9624DD4AED2D}.Release|x64.ActiveCfg = Release|x64
-		{CC13FF8A-9B80-4A09-97CF-9624DD4AED2D}.Release|x64.Build.0 = Release|x64
-		{CC13FF8A-9B80-4A09-97CF-9624DD4AED2D}.Release|x86.ActiveCfg = Release|Win32
-		{CC13FF8A-9B80-4A09-97CF-9624DD4AED2D}.Release|x86.Build.0 = Release|Win32
 		{A0967E25-55C0-465E-86A5-04067A1B7B45}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A0967E25-55C0-465E-86A5-04067A1B7B45}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A0967E25-55C0-465E-86A5-04067A1B7B45}.Debug|x64.ActiveCfg = Debug|Any CPU
