--- cpp/Platform.Collections.Methods.Tests/Platform.Collections.Methods.TemplateLibrary.Tests.cpp
@@ -1,17 +1,0 @@
-#include "pch.h"
-#include "CppUnitTest.h"
-
-using namespace Microsoft::VisualStudio::CppUnitTestFramework;
-
-namespace Platform::Collections::Methods::Tests
-{
-	TEST_CLASS(PlatformCollectionsMethodsTemplateLibraryTests)
-	{
-	public:
-		
-		TEST_METHOD(TestMethod1)
-		{
-
-		}
-	};
-}
